Rename [dirname]/index.[ext] to [dirname].[ext]
For example:
misc/drupal/index.js -> misc/drupal.js

This addresses one downside with the webikon.com approach, which is that all
files, not just HTML files, are renamed to index.[ext]. This corrects that by
restoring the correct file names and updating all references to those files.

This change isn't strictly necessary, but it is nice for the files to have
their intended names, especially because some of those files (like PDF
documents) are accessed directly by the user.